+++ griptape/utils/chat.py
@@ -4,6 +4,8 @@ import logging
 from typing import TYPE_CHECKING, Callable, Optional
 
 from attrs import Factory, define, field
+from rich import print as rprint
+from rich.prompt import Prompt
 
 from griptape.utils.stream import Stream
 
@@ -13,18 +15,27 @@ if TYPE_CHECKING:
 
 @define
 class Chat:
+    class ChatPrompt(Prompt):
+        prompt_suffix = ""  # We don't want rich's default prompt suffix
+
     structure: Structure = field()
     exit_keywords: list[str] = field(default=["exit"], kw_only=True)
     exiting_text: str = field(default="Exiting...", kw_only=True)
     processing_text: str = field(default="Thinking...", kw_only=True)
     intro_text: Optional[str] = field(default=None, kw_only=True)
     prompt_prefix: str = field(default="User: ", kw_only=True)
-    response_prefix: str = field(default="Assistant: ", kw_only=True)
+    response_prefix: str = field(default="Assistant", kw_only=True)
+    input_fn: Callable[[str], str] = field(
+        default=Factory(lambda self: self.default_input_fn, takes_self=True), kw_only=True
+    )
     output_fn: Callable[[str], None] = field(
         default=Factory(lambda self: self.default_output_fn, takes_self=True),
         kw_only=True,
     )
     logger_level: int = field(default=logging.ERROR, kw_only=True)
+
+    def default_input_fn(self, prompt_prefix: str) -> str:
+        return Chat.ChatPrompt.ask(prompt_prefix)
 
     def default_output_fn(self, text: str) -> None:
         from griptape.tasks.prompt_task import PromptTask
@@ -33,9 +44,9 @@ class Chat:
             task for task in self.structure.tasks if isinstance(task, PromptTask) and task.prompt_driver.stream
         ]
         if streaming_tasks:
-            print(text, end="", flush=True)  # noqa: T201
+            rprint(text, end="", flush=True)
         else:
-            print(text)  # noqa: T201
+            rprint(text)
 
     def start(self) -> None:
         from griptape.configs import Defaults
@@ -47,7 +58,7 @@ class Chat:
         if self.intro_text:
             self.output_fn(self.intro_text)
         while True:
-            question = input(self.prompt_prefix)
+            question = self.input_fn(self.prompt_prefix)
 
             if question.lower() in self.exit_keywords:
                 self.output_fn(self.exiting_text)